refactors ping methods into action.py
